Log hosts' statuses (including status info) while waiting for the status of the cluster to update Sometimes a host has issues and blocks the installation, but this information is not logged DNS errors, for instance, prevent the cluster from becoming ready
